correction annomalie de suppression d'entités

+++ app/scripts/controllers/categories.js
@@ -22,16 +22,23 @@ angular.module('inf6150FrontendApp')
   			if (categorie.id){
   				CategoriesService.update({ id: categorie.id }, categorie);
   			}else{
-  				CategoriesService.create(categorie);
+  				CategoriesService.create(categorie, function(){
+            $scope.categories = CategoriesService.findAll();
+          });
   			}  			
   		};
 
-	    $scope.deleteCategories = function(categoriesId) {
-	        CategoriesService.delete({ id: categoriesId })
-	            .$promise.then(
-	                    function(){
-	                        $scope.categories = CategoriesService.findAll();
-	                    }
-	                );
+	    $scope.deleteCategories = function(categoriesId, idx) {
+	        if (categoriesId){
+              CategoriesService.delete({ id: categoriesId })
+              .$promise.then(
+                      function(){
+                          $scope.categories = CategoriesService.findAll();
+                      }
+                  );
+          }else{
+              $scope.categories.splice(idx);
+          }
+          
 	    };
 }]);
